Initial commit
Includes a backup copy of the whole of the TFT_eSPI library
@ version 2.5.23 which I needed because at some point I got shunted to
2.5.43 and things broke. I dumped a copy of this directly on top of the
newer library because I could not (at the time) work out to force
platformIo to use an older version. It may be as easy as editing the
entry in platformio.ini but I have yet to try it.

The code is in a bit of a mess at the moment.